enhancement: handles drag-and-update scenarios
This commit adds ability to support expected behaviour when user
makes changes to calendar by clicking and dragging events around.

The pivotal change here is the resetting of throttler offsets.
While, we want the bulb to turn off immediately when an ongoing
event is moved(dragged) up or below, we also want to refrain from
making very frequent requests to the hue-agent.

Essentially, we want to be swift the first time - when we are
turning the bulb off/on in response to a "drag-update", however
after that, we want to notify hue-agent only at long intervals.

Resetting the throttler offsets allow us to do just that.